refactor delete article

+++ pkg/model/article/article.go
@@ -7,11 +7,11 @@ import (
 
 // Article model
 type Article struct {
-	ID    int
+	ID    int64
 	Title string
 	Body  string
 }
 
 func (a Article) Link() string {
-	return route.Name2URL("articles.show", "id", strconv.FormatInt(int64(a.ID), 10))
+	return route.Name2URL("articles.show", "id", strconv.FormatInt(a.ID, 10))
 }
